Create SAD and ATT banner vs popover experiment (#125)
Task/Issue URL: https://app.asana.com/0/45878998844068/1209518320481653
Tech Design URL:
CC:

### Description
**SAD** =  Set As Default
**ATT** = Add To Taskbar/Dock

The following work adds a one-time popover or banner that we will show
to users. The plan is to measure which is more effective in adding to
the dock and setting it as default.

Given that we want to measure the impact of the popover, banner, or
nothing (control), we will fire experiment pixels when one of these
actions occurs.

Some notes:
- The popover should only be shown in one window, while the banner
should be shown in all windows (and if it is closed in one, it should be
closed in all windows at the same time)
- Neither of the options should be shown again if they are dismissed.
The local flag override marks the prompt as not shown for testing
purposes so we can test it.
- To be eligible for the experiment, you need to have completed
onboarding, not be a new user (more than a week after the first launch),
and not have the browser in the dock or as the default browser.

**Screenshots**


![image](https://github.com/user-attachments/assets/6668869b-ea2a-4d9a-bddd-2ee939bb9e28)

![image](https://github.com/user-attachments/assets/dfb6a6ad-e4a4-4d4a-8de6-9cd9f265748f)

![image](https://github.com/user-attachments/assets/28c9eeca-580a-48aa-8f32-bc8ea596aea1)

![image](https://github.com/user-attachments/assets/e18fdb93-dfdb-4888-9dfb-eadcf38ab589)

### Testing Steps

There are two ways of testing this: 

1. The first one is using the local overrides and selecting either the
popover or banner. To use it, go to the Debug Menu -> Feature Flag
Overrides and select the `popoverVsBanner` experiment.


![image](https://github.com/user-attachments/assets/d6c38712-8533-4b28-9904-6b349e6311a4)

3. The second one uses the following JSON file:
https://www.jsonblob.com/api/1344683245740154880. That adds the
experiment to test:

```json
"setAsDefaultAndAddToDock": {
      "state": "enabled",
      "exceptions": [],
      "features": {
        "popoverVsBannerExperiment": {
          "state": "enabled",
          "cohorts": [
            {
              "name": "control",
              "weight": 1
            },
            {
              "name": "popover",
              "weight": 1
            },
            {
              "name": "banner",
              "weight": 1
            }
          ]
        }
      },
      "hash": "0fe5af5a94e036bf5d78076df6dd771b"
    }
```

You must manually override the Privacy Config with the JSON Blob URL,
which should enroll you in the feature. This one is more random, given
that for testing a particular cohort, you will need to change the weight
in the JSON.

To replace the URL tap `Set custom configuration URL` in the debug menu
above:

![Screenshot 2025-02-28 at 3 28
29 PM](https://github.com/user-attachments/assets/979195ec-6f1a-4033-8268-441bf30ebfb0)

### Impact and Risks
**Low**: This will be handled by a feature flag, so if there is any
problem or complaint, we can just kill it. It is also a popover and a
banner that can be dismissed.

#### What could go wrong?
- The feature is not working as expected and is damaging retention;
Solution -> We can remove the feature flag.
- Showing the prompt more than once. Solution -> We are storing a flag,
so we don’t do this.

### Quality Considerations
I’ve added experiment pixels. No performance issues should arise here.

### Notes to Reviewer
Focus on any edge cases you might have when the prompts are being shown.