AuthV2 \ macOS auth v1 and v2 (#120)
Task/Issue URL:
https://app.asana.com/0/1205842942115003/1209170372758739/f
CC: @miasma13 @diegoreymendez 

### Description

This PR implements the use of AuthV2 in the macOS app.

**The good:**

- AuthV2 is fully integrated into the iOS app.
- The authentication still uses V1 but V2 is configured and added behind
an informal feature flag. (see `isAuthV2Enabled`)

**The bad**

- We now have some code duplication, especially in the settings UI, all
will be removed alongside AuthV1
- "bridging" protocols (see SubscriptionAuthV1toV2Bridge) used where
authv1 and authv2 need to coexist
- The code is not very clean, I prioritised risk reduction and changes
minimisation over code cleanness, all will be improved when we remove
AuthV1

### Testing Steps

All tests need to be executed with App Store build and DMG build, both
using App Store and Stripe purchase flows

1. Purchase subscription
2. Enable VPN
3. Disable VPN
4. Enable VPN
5. Remove subscription
6. Restore subscription
7. Kill app and enable/disable VPN from the system menu

In general, Subscription should work as usual, no user-facing changes
are expected.

### Note to the reviewer

During development and testing, I had a lot of issues with the VPN not
being able to restart immediately after being stopped, I worked 3 days
on that problem until @diegoreymendez tested my code on his machine and
found no issues. If the problem can be replicated during testing (see
steps 2,3,4) then we'll need to fix it, if not, then it's just my
machine misbehaving.

### Quality Considerations

Most of this code is temporary and will be removed around the 24th of
April when AuthV1 will be removed