Privacy Pro Onboarding Experiment - Remove Redundant Code & Update Test (#280)
Task/Issue URL: https://app.asana.com/0/0/1209678164669763

### Description
This PR contains a minor change to remove unnecessary code relating to
the Privacy Pro Onboarding experiment. This code checked privacy pro
eligibility to ensure a user is eligible before enrolling them to the
experiment. However, we have decided to only target the US, and so this
check is redundant. See the related Privacy Config
[here](https://github.com/duckduckgo/privacy-configuration/pull/2876).