Fix logic when moving to previously active tab (#333)
Task/Issue URL:
https://app.asana.com/0/1204006570077678/1209785407607138/f
Tech Design URL:
CC:

### Description
This change fixes two things:
- With the previous change, I introduced a bug in which if the closed
tab was not active, we applied the logic as if it was active, which was
wrong and unintentional.
- The rule of returning to the previous active tab was missing some
logic. On Windows, the previously active tab is only remembered until
the user moves to an existing tab or creates new tabs.

### Testing Steps

**Testing steps of the first bug**
1. Open some site like Wikipedia, and open some links from there using
CMD + Tap
2. Open a new tab and keep that tab selected
3. Right tap one of the child tabs and close it
4. The active tab should be the new tab and no other sibiling or parent
of the closed tab should be selected

**Test we return to the previous active tab**
1. Open five tabs (no related)
2. Select the first tab
3. Use CMD + T or tapping the + in the tab bar to open a new tab
4. Close that tab
5. The first tab should be selected

**Test we do not return to the previous active tab if another tab was
selected**
1. Open five tabs (no related)
2. Select the first tab
3. Use CMD + T or tapping the + in the tab bar to open a new tab
4. Select tab number 5 (the one before the recently opened)
5. Go back to the last tab and close it
6. Tab number 5 should be selected instead of the first tab

### Impact and Risks
Medium: Could disrupt specific features or user flows

#### What could go wrong?
The scenario disrupts other flows when closing tabs.

### Quality Considerations
I’ve added unit tests and UI tests to cover this new scenarios.

### Notes to Reviewer
Do as much test as you can for your user flows.