iOS: Privacy Pro Onboarding Promo - Update Embedded Config To Enable Experiment (#339)
Task/Issue URL:
https://app.asana.com/0/1206488453854252/1208681881503648/f

### Description
We want to enable an experiment Monday March 31st. We can do this
remotely of course. However, the experiment is part of onboarding and so
I want to mitigate any risk of the remote file not being available, and
so enable it also in the embedded config.