Increase navigation bar size and change paddings for new visual refresh (#426)
Task/Issue URL:
https://app.asana.com/0/1204006570077678/1209350564013383/f
Tech Design URL:
CC:

### Description
This adds the new height for the navigation bar; now, it is `52` for
both the default and home page. The pop-up remains unchanged.

It also changes the paddings, and now we have `6` for both the top and
bottom.

This change introduces how we will use the feature flag to apply the
different changes.

### Testing Steps
1. Select the feature flag from `Debug -> Feature Flag Overrides ->
visualRefresh`
2. Open a new window
3. The navigation bar should be taller compared to the other one
4. Turn off the override
5. Open another window
6. The navigation bar size should be back to production values


### Impact and Risks
Low: Minor visual changes, small bug fixes, improvement to existing
features


#### What could go wrong?
I moved some logic around in the `AddressBarSizeClass`. Now, part of
that logic for calculating the address bar height and padding is part of
the `VisualStyleConfigurable`.

### Quality Considerations
Nothing to consider.

### Notes to Reviewer
It’s a simple change. The Dax logo will be larger when the address bar
height increases, but we will move the logo to inside the address bar in
the next PR, so I’m not fixing that now.

—
###### Internal references:
[Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f) |
[Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552) |
[Tech Design
Template](https://app.asana.com/0/59792373528535/184709971311943)